Príloha A2/ocr_server/services/api/requirements.txt
fastapi==0.57.0
gunicorn==20.0.4
uvicorn==0.11.5
prometheusrock==0.1.3
hypercorn
authlib
python-multipart
python-jose-cryptodome
python-dotenv
aniso8601==8.0.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
google-api-core==1.16.0
google-auth==1.10.2
google-cloud-vision==0.41.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
pdf2image==1.11.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
pyOpenSSL==19.1.0
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.0